T#779: Extract 34 forum/thread/message/reactions routes to src/forum/routes.ts
Phase 2.1 of Backend Modularization (Library #102). Mechanical move,
no logic changes.

**Moved to src/forum/routes.ts (~897L)**:
- 34 route handlers covering /api/forum/*, /api/threads, /api/thread/:id*,
  /api/message/:id*, /api/reactions/* path patterns
- All handler bodies verbatim — auth gates, validation, guest-safety
  scanning, mention parsing, search-index upsert all preserved

**Exports**:
- registerForumRoutes(app, sqlite, { hasSessionAuth, requireBeastIdentity,
  isTrustedRequest, wsBroadcast, withRetry, getSupportedEmoji })

**Imports** (cross-domain):
- Role from ../server/rbac.ts
- checkGuestPostRate, checkGuestContentLength, scanForInjection from ../server/guest-safety.ts
- logSecurityEvent from ../server/security-logger.ts
- searchIndexUpsert from ../search/routes.ts (write paths re-index)
- handleThreadMessage, getFullThread, updateThreadStatus from ./handler.ts
- Dynamic-import path fix: ./forum/mentions.ts → ./mentions.ts (same-dir now)

**Pattern**:
- Helpers DI passes withRetry + getSupportedEmoji (utility refs in server.ts)
- SUPPORTED_EMOJI initialized at register-time via getSupportedEmoji()
  helper call, kept as closure-local state (reactions routes mutate this)

**Server.ts changes**:
- Added import after forge import (L105)
- Removed 34 handler bodies from L3542-5014 range (897 lines total)
- Added registerForumRoutes call after registerForgeRoutes (L5442)
- Net: server.ts down ~897 lines (was 6,705, now 5,808)

**Verification**:
- Boot test: server starts clean. All 34 forum routes register cleanly.
- tsc --noEmit: zero new errors (pre-existing TS2322 SecurityEventType /
  Role string-literal-narrow errors unrelated to this PR — present
  before extraction)
- No logic changes — 1:1 mechanical move

**Cumulative T#771-T#779**: server.ts -6,200 lines (11,963 → 5,808, -52%)
in 8 phases over ~3h since 10:23 BKK kickoff.
